+++ R/string.R
@@ -1,0 +1,56 @@
+
+#' Encode a `ps_handle` as a short string
+#'
+#' A convenient format for passing between processes, naming semaphores, or
+#' using as a directory/file name. Will always be 14 alphanumeric characters,
+#' with the first and last characters guarantied to be letters. Encodes the
+#' pid and creation time for a process.
+#'
+#' @param p Process handle.
+#'
+#' @return A process string (scalar character), that can be passed to
+#' `ps_handle()` in place of a pid.
+#'
+#' @export
+#' @examplesIf ps::ps_is_supported() && ! ps:::is_cran_check()
+#' (p <- ps_handle())
+#' (str <- ps_string(p))
+#' ps_handle(pid = str)
+
+ps_string <- function (p = ps_handle()) {
+
+  assert_ps_handle(p)
+
+  process_id <- ps_pid(p)
+  ctime      <- ps_create_time(p)
+  whole_secs <- as.integer(ctime)
+  micro_secs <- as.integer(strsplit(format(ctime, "%OS6"), '.', TRUE)[[1]][[2]])
+
+  # Assumptions:
+  #   ctime between Jan 1st 1970 and Dec 5th 3769.
+  #   max ctime precision = 1/1,000,000 of a second.
+  #   pid <= 7,311,615 (current std max = 4,194,304).
+
+  map <- c(letters, LETTERS, 0:9)
+
+  paste(collapse = '', map[1 + c(
+    floor(process_id / 52 ^ (0:3)) %% 52,
+    floor(whole_secs / 62 ^ (0:5)) %% 62,
+    floor(micro_secs / 52 ^ (0:3)) %% 52 )])
+}
+
+
+ps__str_decode <- function (str) {
+
+  map <- structure(0:61, names = c(letters, LETTERS, 0:9))
+  val <- map[strsplit(str, '', fixed = TRUE)[[1]]]
+
+  process_id <- sum(val[01:04] * 52 ^ (0:3))
+  whole_secs <- sum(val[05:10] * 62 ^ (0:5))
+  micro_secs <- sum(val[11:14] * 52 ^ (0:3))
+
+  time <- whole_secs + (micro_secs / 1000000)
+  time <- as.POSIXct(time, 'GMT')
+
+  ps_handle(pid = process_id, time = time)
+}